schema takes from env and validations

+++ backend/migrating/v2/management/commands/drop_v2_schema.py
@@ -7,6 +7,8 @@ class Command(BaseCommand):
     help = "Drop v2 schema if it exists"
 
     def handle(self, *args, **kwargs):
+        if not V2.SCHEMA_NAME:
+            raise ValueError("SCHEMA_NAME is not defined.")
         with connection.cursor() as cursor:
             cursor.execute(f"DROP SCHEMA IF EXISTS {V2.SCHEMA_NAME} CASCADE")
         self.stdout.write(
